Revert "move KafkaFilter interface to filters package"
This reverts commit 20ac0c094a63472a5c9843914cb86245ba0310e2.
